Split input modules into separate object files
