+++ input/alsa.h
@@ -1,3 +1,5 @@
 // header file for alsa, part of cava.
 
+#pragma once
+
 void *input_alsa(void *data);
