+++ input/fifo.h
@@ -1,3 +1,5 @@
 // header files for fifo, part of cava
+
+#pragma once
+
 void *input_fifo(void *data);
-int write_to_fftw_input_buffers(int16_t buf[], int16_t frames, void *data);
