+++ input/portaudio.h
@@ -1,1 +1,3 @@
+#pragma once
+
 void *input_portaudio(void *audiodata);
