+++ input/shmem.h
@@ -1,0 +1,5 @@
+// header file for shmem, part of cava.
+
+#pragma once
+
+void *input_shmem(void *data);
