+++ input/sndio.h
@@ -1,0 +1,5 @@
+// header file for sndio, part of cava.
+
+#pragma once
+
+void *input_sndio(void *data);
